% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{factor_palette}
\alias{factor_palette}
\title{Create a named color palette for a factor}
\usage{
factor_palette(x, colors, ...)
}
\arguments{
\item{x}{A factor or character vector to use for the levels}

\item{colors}{A vector of colors}

\item{...}{Here, the user can provide numeric index values for colors that
should be used first, second, third, and so on in the palette. This allows
for the colors to be rearranged if they are not quite cutting it in the
default order.}
}
\value{
A named vector of colors corresponding to the factor.
}
\description{
Create a named color palette for a factor
}
\details{
This function may be most useful in conjunction with \code{dput()},
  which will print out the code for the resultant palette. This way, color
  palettes can be hard-coded into scripts, which will preserve the color
  scheme even if the levels of the factor were to change in the future.
}
