% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheat.R
\name{pheat}
\alias{pheat}
\alias{pheat.matrix}
\alias{pheat.data.frame}
\title{Build a pheatmap object in stages}
\usage{
pheat(x, ...)

\method{pheat}{matrix}(x)

\method{pheat}{data.frame}(x, rownames_in = 1)
}
\arguments{
\item{x}{A matrix or data frame of data for the heatmap}

\item{...}{Additional arguments passed to other methods}

\item{rownames_in}{The column of the data frame that gives the rownames
for the heatmap. Can be specified as an integer index or with the column
name in quotes. Set to \code{NULL} to use the existing rownames of the
data frame.}
}
\description{
Build a pheatmap object in stages
}
\examples{
pheat(vendor_props)

# Re-format heatmap data in a data frame
vendor_props_df <- data.frame(
  taxon_id = rownames(vendor_props),
  vendor_props[,1:15])
rownames(vendor_props_df) <- NULL
pheat(vendor_props_df, rownames_in = "taxon_id")
}
