% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{pheat_annotate}
\alias{pheat_annotate}
\alias{pheat_annotate_rows}
\alias{pheat_annotate_cols}
\title{Annotate rows and columns of the heatmap}
\usage{
pheat_annotate_rows(ph, annotation, rownames_in = 1)

pheat_annotate_cols(ph, annotation, rownames_in = 1)
}
\arguments{
\item{ph}{Heatmap of class "pheat"}

\item{annotation}{A data frame of annotations}

\item{rownames_in}{The column in the annotation data frame that corresponds
to the rows or columns of the data matrix for the heatmap. Set to
\code{NULL} to use the existing rownames of the annotation data frame.}
}
\description{
Annotate rows and columns of the heatmap
}
\examples{
ph <- pheat(vendor_props)
pheat_annotate_rows(ph, vendor_taxa)
pheat_annotate_cols(ph, vendor_samples)
}
