% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{pheat_annotation_color}
\alias{pheat_annotation_color}
\title{Set color palettes for heatmap annotations}
\usage{
pheat_annotation_color(ph, ...)
}
\arguments{
\item{...}{Expects keyword arguments where the keyword is the column name
for the annotation, and the value is a (named) character vector of colors}
}
\value{
A modified heatmap of class "pheat"
}
\description{
Set color palettes for heatmap annotations
}
