% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{pheat_annotation_legend}
\alias{pheat_annotation_legend}
\title{Adjust the legend for annotations in the heatmap}
\usage{
pheat_annotation_legend(
  ph,
  show = TRUE,
  show_names_row = TRUE,
  show_names_col = TRUE
)
}
\arguments{
\item{ph}{Heatmap of class "pheat"}

\item{show}{If \code{FALSE}, don't show the legends for annotations}

\item{show_names_row, show_names_col}{If \code{FALSE}, don't show names for
row or column annotation tracks}
}
\value{
A modified heatmap of class "pheat"
}
\description{
Adjust the legend for annotations in the heatmap
}
\examples{
phc <- pheat_annotate_cols(pheat(vendor_props), vendor_samples)
pheat_annotation_legend(phc, show_names_col = FALSE)
}
