% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{pheat_color}
\alias{pheat_color}
\title{Heatmap color scale}
\usage{
pheat_color(ph, color = NA, breaks = NA, na_color = "#DDDDDD")
}
\arguments{
\item{ph}{Heatmap of class "pheat"}

\item{color}{Color palette for hetamap}

\item{breaks}{Breakpoints between colors in the heatmap. Should be a numeric
vector that is one element longer than the color palette}

\item{na_color}{Color for \code{NA} values in the heatmap}
}
\value{
A modified heatmap of class "pheat"
}
\description{
Heatmap color scale
}
