% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{pheat_color_saturated}
\alias{pheat_color_saturated}
\title{Use a saturated rainbow color palette for the heatmap}
\usage{
pheat_color_saturated(
  ph,
  saturation_limit = 0.4,
  upper_limit = 1,
  na_color = "#DDDDDD"
)
}
\arguments{
\item{saturation_limit}{The value after which the color stops changing}

\item{upper_limit}{The upper limit of the scale}

\item{na_color}{Color for \code{NA} values in the heatmap}
}
\value{
A modified heatmap of class "pheat"
}
\description{
Use a saturated rainbow color palette for the heatmap
}
