% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{pheat_display}
\alias{pheat_display}
\alias{pheat_display_main}
\alias{pheat_display_rows}
\alias{pheat_display_cols}
\title{Set display style for heatmap}
\usage{
pheat_display_main(
  ph,
  border_color = "grey60",
  fontsize = 10,
  title = NA,
  display_numbers = F
)

pheat_display_rows(
  ph,
  height = 10,
  show_names = TRUE,
  fontsize = NA,
  gaps = NULL,
  labels = NULL,
  dendrogram_width = NA
)

pheat_display_cols(
  ph,
  width = 10,
  show_names = TRUE,
  fontsize = NA,
  gaps = NULL,
  labels = NULL,
  dendrogram_height = NA
)
}
\arguments{
\item{ph}{Heatmap of class "pheat"}

\item{border_color}{Border color for cells. Set to \code{NA} to remove cell
borders.}

\item{fontsize}{Font size for row or column names}

\item{title}{Title of heatmap}

\item{display_numbers}{logical determining if the numeric values are also printed to the cells. If this is a matrix (with same dimensions as original matrix), the contents of the matrix are shown instead of original values.}

\item{height, width}{Row height and column width. If set to \code{NA}, the
size varies according to the available space}

\item{show_names}{Show row or column names}

\item{gaps}{Location of gaps between rows or columns. Gaps are located
\emph{after} the specified locations in the heatmap, moving from top to
bottom for rows and left to right for columns. As a convenience, the user
may provide a vector of row names after which gaps are to be placed.}

\item{labels}{Custom labels to use in place of the row or column names in
the data}

\item{dendrogram_width, dendrogram_height}{Width and height of dendrograms
for rows and columns, respectively, in points. If set to NA, the default
value of 50 points will be used.}
}
\value{
A modified heatmap of class "pheat"
}
\description{
Set display style for heatmap
}
\examples{
ph <- pheat(vendor_props)

# No border
pheat_display_main(ph, border_color = NA)

# Gaps between bacterial phyla
pheat_display_rows(ph, gaps = c(5, 10))

# Re-label samples
pheat_display_cols(ph, labels = paste("Sample", 1:30))
}
