% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{pheat_legend}
\alias{pheat_legend}
\title{Adjust the legend for the heatmap}
\usage{
pheat_legend(ph, breaks = NA, labels = NA, show = TRUE)
}
\arguments{
\item{ph}{Heatmap of class "pheat"}

\item{breaks}{Breakpoints for labels in the heatmap legend}

\item{labels}{Custom labels to use in the heatmap legend}

\item{show}{If \code{FALSE}, don't show the legend}
}
\value{
A modified heatmap of class "pheat"
}
\description{
Adjust the legend for the heatmap
}
