% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{print.pheat}
\alias{print.pheat}
\title{Print the heatmap to the screen}
\usage{
\method{print}{pheat}(ph, ...)
}
\arguments{
\item{ph}{Heatmap of class "pheat"}
}
\description{
Print the heatmap to the screen
}
