% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor.R
\docType{data}
\name{vendor_props}
\alias{vendor_props}
\alias{vendor_taxa}
\alias{vendor_samples}
\title{Microbiome of mouse cecum and feces}
\format{
The data matrix \code{vendor_props} contains the bacterial taxon
  proportions in each sample. Only taxa with >1\% mean relative abundance
  were kept in the matrix.

The associated data frame \code{vendor_taxa} describes the taxa that
  comprise the rows of \code{vendor_props}:
  \describe{
    \item{taxon}{The taxon represented in the data matrix.}
    \item{phylum}{The phylum of the taxon.}
  }

The associated data frame \code{vendor_samples} describes the
  samples comprising the columns of \code{vendor_props}:
  \describe{
    \item{sample_id}{The sample represented in the data matrix.}
    \item{sample_type}{The type of sample, either Cecum or Feces.}
    \item{vendor}{The vendor from which the mouse was purchased.}
  }
}
\usage{
vendor_props

vendor_taxa

vendor_samples
}
\description{
Mice were purchased from four different vendors. Samples from the cecum and
feces were submitted for 16S rRNA marker gene sequencing to profile the
bacterial communities.
}
\keyword{datasets}
